airflow/requirements.txt
openai
pinecone-client
langchain-openai
langchain
langchain-core
bs4
langchain-community
langchain-pinecone
boto3
gitpython
snowflake-connector-python